+++ source/App/source/ExampleHost.WebApi.Tests/Integration/AuthenticationTests.cs
@@ -15,17 +15,22 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Net;
 using System.Security.Claims;
-using System.Text;
 using ExampleHost.WebApi.Tests.Fixtures;
+using ExampleHost.WebApi04.Controllers;
 using FluentAssertions;
+using Microsoft.Identity.Client;
 using Microsoft.IdentityModel.Tokens;
 using Xunit;
 
 namespace ExampleHost.WebApi.Tests.Integration;
 
 /// <summary>
-/// Authentication tests using a standard token (not nested, so this is not like
-/// in production) to verify that tokens are configured to be validated as expected.
+/// Authentication tests using a nested token (a token which contains both an
+/// external and an internal token) to verify that tokens are configured
+/// to be validated as expected.
+///
+/// Similar tests exists for Function App in the 'AuthenticationTests' class
+/// located in the 'ExampleHost.FunctionApp.Tests' project.
 /// </summary>
 [Collection(nameof(AuthenticationHostCollectionFixture))]
 public sealed class AuthenticationTests
@@ -36,6 +41,19 @@ public sealed class AuthenticationTests
     }
 
     private AuthenticationHostFixture Fixture { get; }
+
+    [Fact]
+    public async Task CallingApi04Get_NoEndpoint_Returns404()
+    {
+        // Arrange
+        using var request = new HttpRequestMessage(HttpMethod.Get, "webapi04/authentication/does_not_exist");
+
+        // Act
+        using var actualResponse = await Fixture.Web04HttpClient.SendAsync(request);
+
+        // Assert
+        actualResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
+    }
 
     [Fact]
     public async Task CallingApi04Get_Anonymous_Succeeds()
@@ -55,19 +73,6 @@ public sealed class AuthenticationTests
     }
 
     [Fact]
-    public async Task CallingApi04Get_NoEndpoint_Returns404()
-    {
-        // Arrange
-        using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/does_not_exist");
-
-        // Act
-        using var actualResponse = await Fixture.Web04HttpClient.SendAsync(request);
-
-        // Assert
-        actualResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-    }
-
-    [Fact]
     public async Task CallingApi04Get_AuthRequiredButNoToken_Unauthorized()
     {
         // Arrange
@@ -82,12 +87,28 @@ public sealed class AuthenticationTests
     }
 
     [Fact]
+    public async Task CallingApi04Get_AuthWithFakeToken_Unauthorized()
+    {
+        // Arrange
+        var requestIdentification = Guid.NewGuid().ToString();
+        var authenticationHeader = CreateAuthenticationHeaderWithFakeToken();
+
+        // Act
+        using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/auth/{requestIdentification}");
+        request.Headers.Add("Authorization", authenticationHeader);
+        using var actualResponse = await Fixture.Web04HttpClient.SendAsync(request);
+
+        // Assert
+        actualResponse.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
+    }
+
+    [Fact]
     public async Task CallingApi04Get_AuthWithToken_Allowed()
     {
         // Arrange
         var requestIdentification = Guid.NewGuid().ToString();
         var authenticationResult = await Fixture.GetTokenAsync();
-        var authenticationHeader = authenticationResult.CreateAuthorizationHeader();
+        var authenticationHeader = await CreateAuthenticationHeaderWithNestedTokenAsync(authenticationResult);
 
         // Act
         using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/auth/{requestIdentification}");
@@ -107,7 +128,7 @@ public sealed class AuthenticationTests
         // Arrange
         var requestIdentification = Guid.NewGuid().ToString();
         var authenticationResult = await Fixture.GetTokenAsync();
-        var authenticationHeader = authenticationResult.CreateAuthorizationHeader();
+        var authenticationHeader = await CreateAuthenticationHeaderWithNestedTokenAsync(authenticationResult);
 
         // Act
         using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/auth/{requestIdentification}");
@@ -124,10 +145,10 @@ public sealed class AuthenticationTests
     {
         // Arrange
         var authenticationResult = await Fixture.GetTokenAsync();
-        var authenticationHeader = authenticationResult.CreateAuthorizationHeader();
+        var authenticationHeader = await CreateAuthenticationHeaderWithNestedTokenAsync(authenticationResult);
 
         // Act
-        using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/user");
+        using var request = new HttpRequestMessage(HttpMethod.Get, "webapi04/authentication/user");
         request.Headers.Add("Authorization", authenticationHeader);
         using var actualResponse = await Fixture.Web04HttpClient.SendAsync(request);
 
@@ -138,26 +159,33 @@ public sealed class AuthenticationTests
         Assert.True(Guid.TryParse(content, out _));
     }
 
-    [Fact]
-    public async Task CallingApi04Get_AuthWithFakeToken_Unauthorized()
+    private static string CreateAuthenticationHeaderWithFakeToken()
     {
-        // Arrange
-        var requestIdentification = Guid.NewGuid().ToString();
-
-        var securityKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("not-a-secret-keynot-a-secret-key"));
+        var securityKey = new SymmetricSecurityKey("not-a-secret-keynot-a-secret-key"u8.ToArray());
         var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256);
-        var subClaim = new Claim("sub", Guid.NewGuid().ToString());
+        var userIdAsSubClaim = new Claim("sub", Guid.NewGuid().ToString());
+        var actorIdAsAzpClaim = new Claim("azp", Guid.NewGuid().ToString());
 
-        var securityToken = new JwtSecurityToken(claims: new[] { subClaim }, signingCredentials: credentials);
+        var securityToken = new JwtSecurityToken(claims: [userIdAsSubClaim, actorIdAsAzpClaim], signingCredentials: credentials);
         var tokenHandler = new JwtSecurityTokenHandler();
         var token = tokenHandler.WriteToken(securityToken);
 
-        // Act
-        using var request = new HttpRequestMessage(HttpMethod.Get, $"webapi04/authentication/auth/{requestIdentification}");
-        request.Headers.Add("Authorization", $"Bearer {token}");
-        using var actualResponse = await Fixture.Web04HttpClient.SendAsync(request);
+        var authenticationHeader = $"Bearer {token}";
+        return authenticationHeader;
+    }
 
-        // Assert
-        actualResponse.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
+    /// <summary>
+    /// Calls the <see cref="MockedTokenController"/> to create an "internal token"
+    /// and returns a 'Bearer' authentication header.
+    /// </summary>
+    private async Task<string> CreateAuthenticationHeaderWithNestedTokenAsync(AuthenticationResult externalAuthenticationResult)
+    {
+        using var tokenRequest = new HttpRequestMessage(HttpMethod.Post, "webapi04/token");
+        tokenRequest.Content = new StringContent(externalAuthenticationResult.AccessToken);
+        using var tokenResponse = await Fixture.Web04HttpClient.SendAsync(tokenRequest);
+
+        var nestedToken = await tokenResponse.Content.ReadAsStringAsync();
+        var authenticationHeader = $"Bearer {nestedToken}";
+        return authenticationHeader;
     }
 }